requirements.txt
asgiref==3.8.1
Django==5.0.4
django-model-utils==4.5.0
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.11.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
